site/requirements.txt
aioredis==1.3.1
appnope==0.1.2
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
boto3==1.18.40
botocore==1.21.40
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dateutils==0.6.12
decorator==5.0.9
Django==3.2.6
django-bootstrap4==3.0.1
django-environ==0.7.0
django-extensions==3.1.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
matplotlib-inline==0.1.2
msgpack==1.0.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.0.5
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
zope.interface==5.4.0
gunicorn==20.0.4
supervisor==4.2.2
pandas